creates all necessary csv files
