+++ preprocessing/preprocessing.py
@@ -1,11 +1,20 @@
 import pandas as pd
+import os
+import sys
+from datetime import datetime, timedelta
 
-from preprocessing.canteen_tail.canteen_tail import add_canteen_history
-from preprocessing.decision_tree.decision_tree_preprocessing import (
-    get_dataset_with_weekday,
+sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../analysis")
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../preprocessing"
 )
-from preprocessing.start_of_year.start_of_year import add_diff_from_start_year
-from preprocessing.weather.categorize_weather import categorize_temperature
+from canteen_tail.canteen_tail import add_canteen_history
+from decision_tree.decision_tree_preprocessing import get_dataset_with_weekday
+from start_of_year.start_of_year import add_diff_from_start_year
+from weather.categorize_weather import categorize_temperature
+from analysis.combined_dataset import get_holiday_data
+from analysis.weather_data.weather_forecast import get_weather_forecast
+from analysis.combined_dataset import create_csv
+from helpers import map_bool_to_int
 
 
 def from_intervall_to_col(df, new_col, old_col):
@@ -13,6 +22,24 @@ def from_intervall_to_col(df, new_col, old_col):
         lambda row: 1.0 if row[old_col] == new_col else 0.0, axis=1
     )
     return df
+
+
+def get_df_next_days():
+    date_today = datetime.now()
+    from_date = (date_today + timedelta(1)).date()
+    to_date = (date_today + timedelta(7)).date()
+
+    weather = get_weather_forecast()
+    holiday = get_holiday_data(from_date, to_date)
+    merged = pd.merge(
+        weather, holiday, left_index=True, right_index=True, how="left"
+    )
+    merged["Canteen"] = -1
+    map_bool_to_int(merged, "holiday")
+    map_bool_to_int(merged, "vacation")
+    map_bool_to_int(merged, "inneklemt")
+    merged.index.name = "date"
+    return merged
 
 
 def preprocess_data(df):
@@ -31,8 +58,6 @@ def preprocess_data(df):
 
 
 def preprocess_for_ml(df):
-    df = preprocess_data(df)
-
     from_intervall_to_col(df, "preferred_work_temp", "avg_temp")
     from_intervall_to_col(df, "stay_home_temp", "avg_temp")
 
@@ -48,19 +73,32 @@ def preprocess_for_ml(df):
     return df
 
 
-def save_dataframes(dataframe):
+def save_dataframes_next_days():
+    df = get_df_next_days()
+    decision_tree_df = preprocess_data(df)
+    ml_df = preprocess_for_ml(decision_tree_df.copy())
+    decision_tree_df.to_csv(
+        index=True, path_or_buf="../data/decision_tree_df_next_days.csv"
+    )
+    ml_df.to_csv(index=True, path_or_buf="../data/ml_df_next_days.csv")
+
+
+def create_and_save_dataframes():
+    create_csv()
+    dataframe = pd.read_csv("../data/dataset.csv", index_col="date")
+
     df = dataframe.copy()
     decision_tree_df = preprocess_data(df)
-    ml_df = preprocess_for_ml(df)
+    ml_df = preprocess_for_ml(decision_tree_df.copy())
     decision_tree_df.to_csv(
         index=True, path_or_buf="../data/decision_tree_df.csv"
     )
     ml_df.to_csv(index=True, path_or_buf="../data/ml_df.csv")
+    save_dataframes_next_days()
 
 
 def main():
-    dataframe = pd.read_csv("../data/dataset.csv", index_col="date")
-    save_dataframes(dataframe)
+    create_and_save_dataframes()
 
 
 if __name__ == "__main__":
